# Conflicts:
#	.github/workflows/build.yml
#	modules/lwjgl/core/src/main/java/org/lwjgl/system/Upcalls.java
#	modules/lwjgl/core25/src/generated/java/org/lwjgl/system/JNI.java

+++ modules/lwjgl/core/src/templates/kotlin/core/linux/liburing/templates/liburing.kt
@@ -1814,14 +1814,14 @@ ENABLE_WARNINGS()""")
         io_uring.p("ring")
     )
 
-    int(
+    ssize_t(
         "mlock_size",
 
         unsigned("entries"),
         unsigned("flags")
     )
 
-    int(
+    ssize_t(
         "mlock_size_params",
 
         unsigned("entries"),
